Add a "default" option to the moused configuration
Although the moused man page does not describe it, the moused code has a "default" setting and it is distinct from the "low" setting. According to the psm driver page, the default value "...selects the default resolution for the device (whatever it is)."

I'm not certain of the effect on every mouse, but every mouse or trackpad or what have you on trueos that I have run on has needed a high resolution mouse. I suspect putting this value at default may just magically lead to a better effect. Plus, I can't think of a workable X resolution where 50dpi (the low setting) is viable.

I've inserted default first, so it gets selected first, so people who never touch this screen should be OK.